… ggml/src/ggml-qnn/npu/CMakeLists.txt
… ggml/src/ggml-qnn/npu/device/op_mul_mat.cpp
… ggml/src/ggml-qnn/npu/device/op_rope.cpp
… ggml/src/ggml-qnn/npu/device/tensor.hpp
… ggml/src/ggml-qnn/npu/device/type_traits.cpp
… ggml/src/ggml-qnn/npu/device/type_traits.hpp
… ggml/src/ggml-qnn/npu/device/util.hpp
… ggml/src/ggml-qnn/npu/device/vec_math.inl
… ggml/src/ggml-qnn/npu/device/vec_ops.hpp
… ggml/src/ggml-qnn/npu/device/vec_ops.inl
… ggml/src/ggml-qnn/npu/host/graph.cpp
… ggml/src/ggml-qnn/npu/host/host_device.cpp
… ggml/src/ggml-qnn/npu/host/tensor.hpp
… ggml/src/ggml-qnn/npu/host/util.cpp
… ggml/src/ggml-qnn/npu/host/util.hpp
… ggml/src/ggml-qnn/npu/idl/hexagon_npu.idl