Related to #34 
Related to #51 

## Overview
- Introduces a lookup-table (LUT) based dequantization path to reduce per-element arithmetic and speed up matmul on quantized weights.
- Applies to low-bit quant formats where code values can be mapped via compact per-block LUTs (e.g., 4/5/8-bit).
- Goal: higher tokens/s and lower latency, with no accuracy regression.

## Key Changes
- Added LUT generation per quant block using existing block metadata (e.g., scale/zero or equivalent).
- Integrated a LUT-based dequant fast path for supported quant formats.
- Safe fallback to legacy dequant when the format is unsupported or the LUT path is disabled.
- Kept hot loops branch-light and memory-access contiguous to help vectorization and caching.
- Added guards/tests to validate parity with the legacy path.

## Performance

Test setup: 8gen2
Test suite: `test-backend-ops`

| Matrix Type | Baseline (0979133ea) | Optimized (38935b67e) | Improvement |
|-------------|-----------------------|------------------------|-------------|
| q4_0        | 16.52 GFLOPS          | 23.54 GFLOPS           | +42.5%      |
| q4_K        | 0.196 GFLOPS          | 2.38 GFLOPS            | 12.2x       |

- q4_K 
  | n   | Baseline (GFLOPS) | Optimized (GFLOPS) | Speedup |
  |-----|--------------------|--------------------|---------|
  | 1   | 0.20               | 2.38               | 12.2x   |
  | 2   | 0.39               | 4.61               | 11.8x   |
  | 3   | 0.58               | 6.69               | 11.5x   |
  | 4   | 0.78               | 8.65               | 11.1x   |
  | 5   | 0.97               | 10.49              | 10.8x   |
  | 8   | 1.53               | 15.39              | 10.1x   |
  | 512 | 35.86              | 56.07              | 1.56x   |

- q4_0 
  | n   | Baseline (GFLOPS) | Optimized (GFLOPS) | Speedup |
  |-----|--------------------|--------------------|---------|
  | 1   | 16.52              | 23.54              | 1.43x   |
  | 2   | 26.82              | 36.15              | 1.35x   |
  | 3   | 33.65              | 43.65              | 1.30x   |
  | 4   | 38.50              | 48.55              | 1.26x   |
  | 5   | 42.16              | 52.00              | 1.23x   |
  | 8   | 48.65              | 57.67              | 1.19x   |
  | 512 | 55.85              | 58.62              | 1.05x   |

Notes
- Baseline log: [test-backend-ops-perf-all.release.hexagon.0979133ea.log](https://drive.google.com/file/d/11PzMaZVJ-rUBhY-YzKM6Qs6PwEjEZuKi/view?usp=sharing)
- Optimized log: [test-backend-ops-perf-all.release.hexagon.38935b67e.log](https://drive.google.com/file/d/12k3RRFkn96yy1uXOJ-13GN7wX7LxyxXJ/view?usp=sharing)

## Unit tests

Test setup: 8gen2
Test suite: `test-backend-ops`

```log
Backend 2/2: CPU
  Skipping
2/2 backends passed
[1;32mOK[0m
```

[test-backend-ops-all.release.hexagon.38935b67e.7z](https://drive.google.com/file/d/1sCVdnexbaedWP4rTPTWC10hrSPJmriwv/view?usp=sharing)